+++ src/plugins/rules.js
@@ -1,16 +1,16 @@
 /* @flow */
-import SimpleRule from './SimpleRule'
-import KeyframeRule from './KeyframeRule'
-import ConditionalRule from './ConditionalRule'
-import FontFaceRule from './FontFaceRule'
-import ViewportRule from './ViewportRule'
+import SimpleRule from '../rules/SimpleRule'
+import KeyframesRule from '../rules/KeyframesRule'
+import ConditionalRule from '../rules/ConditionalRule'
+import FontFaceRule from '../rules/FontFaceRule'
+import ViewportRule from '../rules/ViewportRule'
 import type {Plugin, RuleOptions, Rule, JssStyle} from '../types'
 
 const classes = {
   '@charset': SimpleRule,
   '@import': SimpleRule,
   '@namespace': SimpleRule,
-  '@keyframes': KeyframeRule,
+  '@keyframes': KeyframesRule,
   '@media': ConditionalRule,
   '@supports': ConditionalRule,
   '@font-face': FontFaceRule,