+++ tests/integration/hooks.js
@@ -2,14 +2,14 @@ import expect from 'expect.js'
 import {stripIndent} from 'common-tags'
 import {create} from '../../src'
 import StyleSheet from '../../src/StyleSheet'
-import {generateClassName} from '../utils'
+import {createGenerateClassName} from '../utils'
 import PluginsRegistry from '../../src/PluginsRegistry'
 
 describe('Integration: hooks', () => {
   let jss
 
   beforeEach(() => {
-    jss = create({generateClassName})
+    jss = create({createGenerateClassName})
   })
 
   describe('common', () => {
@@ -33,16 +33,16 @@ describe('Integration: hooks', () => {
     it('should call hooks in the correct order', () => {
       jss.use({
         onProcessRule: (rule) => {
-          if (rule.name === 'a') {
+          if (rule.key === 'a') {
             rule.options.sheet.addRule('b', {color: 'green'}, {index: 1})
           }
         }
       })
 
-      const classNames = []
+      const selectors = []
       jss.use({
         onProcessRule: (rule) => {
-          classNames.push(rule.className)
+          selectors.push(rule.selector)
         }
       })
 
@@ -65,7 +65,7 @@ describe('Integration: hooks', () => {
         '  color: blue;\n' +
         '}'
       )
-      expect(classNames).to.eql(['b-id', 'a-id', 'c-id'])
+      expect(selectors).to.eql(['.b-id', '.a-id', '.c-id'])
     })
 
     it('should warn when unknown hook name is used', () => {
@@ -348,7 +348,7 @@ describe('Integration: hooks', () => {
       })
 
       expect(receivedStyle).to.be(newStyle)
-      expect(receivedRule.type).to.be('regular')
+      expect(receivedRule.type).to.be('style')
       expect(receivedSheet).to.be(sheet)
       expect(localExecuted).to.be(1)
     })